Internalise elmslie-android module
