+++ elmslie-android/src/main/java/vivid/money/elmslie/android/logger/DefaultInitContentProvider.kt
@@ -1,38 +1,21 @@
 package vivid.money.elmslie.android.logger
 
-import android.content.ContentProvider
-import android.content.ContentValues
+import android.app.Application
 import android.content.pm.ApplicationInfo
-import android.database.Cursor
-import android.net.Uri
+import vivid.money.elmslie.android.processdeath.ProcessDeathDetector
 import vivid.money.elmslie.core.config.ElmslieConfig
 
-class DefaultInitContentProvider : ContentProvider() {
+internal class DefaultInitContentProvider : EmptyContentProvider() {
 
-    override fun onCreate(): Boolean {
-        val isDebug = 0 != context!!.applicationInfo.flags and ApplicationInfo.FLAG_DEBUGGABLE
-        if (isDebug) ElmslieConfig.defaultDebugLogger() else ElmslieConfig.defaultReleaseLogger()
-        return true
+    override fun init() {
+        initDefaultLogger()
+        initProcessDeathDetector()
     }
 
-    override fun query(
-        uri: Uri,
-        projection: Array<out String>?,
-        selection: String?,
-        selectionArgs: Array<out String>?,
-        sortOrder: String?
-    ): Cursor = error("not allowed")
+    private fun initProcessDeathDetector() = ProcessDeathDetector.init(context!!.applicationContext as Application)
 
-    override fun getType(uri: Uri): String = error("not allowed")
-
-    override fun insert(uri: Uri, values: ContentValues?): Uri = error("not allowed")
-
-    override fun delete(uri: Uri, selection: String?, selectionArgs: Array<out String>?): Int = error("not allowed")
-
-    override fun update(
-        uri: Uri,
-        values: ContentValues?,
-        selection: String?,
-        selectionArgs: Array<out String>?
-    ): Int = error("not allowed")
+    private fun initDefaultLogger() {
+        val isDebug = 0 != context!!.applicationInfo.flags and ApplicationInfo.FLAG_DEBUGGABLE
+        if (isDebug) ElmslieConfig.defaultDebugLogger() else ElmslieConfig.defaultReleaseLogger()
+    }
 }
