+++ elmslie-core/src/main/java/vivid/money/elmslie/core/store/Option.kt
@@ -3,14 +3,4 @@ package vivid.money.elmslie.core.store
 /**
  * Use this wrapper in case when expected not null value but null value can be present
  */
-internal data class Option<T : Any>(
-    val value: T?
-) {
-
-    val isEmpty = value == null
-
-    companion object {
-
-        fun <T : Any> none() = Option<T>(null)
-    }
-}
+internal class Option<T : Any>(val value: T?)
